backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyPDF2
openai
python-dotenv
python-docx

pandas
openpyxl
xlrd
